support adding Programs together, refactor py Program into its own module
